Refactor : header 우측 컴포넌트 분리
